requirements.txt
python-dotenv
openai
pinecone
pandas
numpy
scikit-learn
python-docx
PyPDF2
langchain
PyMuPDF
